Fix 'active' command post Slack API change
Slack has removed the "presence" option from the users API; this means
we can no longer fetch the status of the users in one operation, but
must instead  fetch all the users, then inquire for each one whether or
not the user is active. This has to be slowed down some because too many
checks at once causes Slack to rate-limit the API.

The code now just dumps out the users as they are found instead of
putting them on one line for simplicity.